Edit is working and refactored
