+++ lib/index.js
@@ -6,3 +6,4 @@ require('./ajax-requests/postFood')
 require('./ajax-responses/foodResponses')
 require('./event-listeners/foodsListeners')
 require('./objects/foodsTableFilter')
+require('./ajax-requests/editFoods')
